requirements.txt
# Kalshi Trading Bot Dependencies
# ================================

# HTTP client (async)
httpx>=0.27.0

# WebSocket support
websockets>=12.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Cryptography for RSA signing
cryptography>=42.0.0

# Type hints (Python 3.9+ backport)
typing-extensions>=4.0.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
